requirements.txt
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiosignal==1.3.1
altair==4.0.0
anyio==3.7.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.2.0
blinker==1.8.2
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.5.14
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
entrypoints==0.4
exceptiongroup==1.2.2
faiss-cpu==1.7.4
fastapi==0.112.0
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.2
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.28
langchain-text-splitters==0.2.2
langsmith==0.1.96
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
numexpr==2.10.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.18.1
openai==0.28.1
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pinecone-client==2.2.1
posthog==3.5.0
protobuf==3.20.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==1.10.17
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.1
pyparsing==3.1.2
pypdf==4.3.1
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
semver==3.0.2
sentence-transformers==3.0.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
st-annotated-text==4.0.1
starlette==0.37.2
streamlit==1.18.1
streamlit-camera-input-live==0.2.0
streamlit-embedcode==0.1.2
streamlit-extras==0.1.5
streamlit-keyup==0.2.4
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.4.0
tornado==6.4.1
tqdm==4.66.4
transformers==4.43.3
triton==3.0.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0
validators==0.33.0
watchdog==4.0.1
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
